% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedMask.R
\name{DelayedMask}
\alias{DelayedMask}
\alias{DelayedMask-class}
\alias{dim,DelayedMask-method}
\alias{dimnames,DelayedMask-method}
\alias{chunkdim,DelayedMask-method}
\alias{path,DelayedMask-method}
\alias{is_sparse,DelayedMask-method}
\alias{extract_array,DelayedMask-method}
\alias{extract_sparse_array,DelayedMask-method}
\alias{OLD_extract_sparse_array,DelayedMask-method}
\title{Delayed masking}
\usage{
DelayedMask(x, placeholder, force = FALSE)
}
\arguments{
\item{x}{An existing \pkg{DelayedArray} seed.}

\item{placeholder}{Placeholder value to replace with \code{NA}.
This should be of the same type as \code{\link{type}(x)}.}

\item{force}{Whether to forcibly create a DelayedMask if \code{placeholder} is already \code{NA}.}
}
\value{
A DelayedMask object, to be wrapped in a \code{\link{DelayedArray}}.
If \code{force=FALSE} and \code{placeholder} is already \code{NA}, \code{x} is directly returned.
}
\description{
Delayed masking of missing values, based on replacement of placeholder values with \code{NA}.
This allows missingness to be encoded in frameworks without the same concept of \code{NA} as R.
}
\examples{
original <- DelayedArray(matrix(rpois(40, lambda=2), ncol=5))
original
masked <- DelayedMask(original, 0)
DelayedArray(masked)

}
\author{
Aaron Lun
}
