% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DelayedNoOp}
\alias{DelayedNoOp}
\alias{DelayedNoOp-class}
\title{Delayed no-op base class}
\description{
A DelayedArray layer that does not modify the contents of its seed.
Generics like \code{\link[DelayedArray]{extract_array}} simply call the corresponding method for the seed.
DelayedNoOp subclasses are used by \pkg{alabaster.matrix} to provide metadata that persists throughout delayed operations,
see \linkS4class{ReloadedArraySeed} and \linkS4class{ToBeRealizedArraySeed} for more details.
}
\details{
We expose this subclass to facilitate traversal of a DelayedArray's operation tree that contains \pkg{alabaster.matrix} classes.
Specifically, these no-op instances can be easily detected and skipped without having to list each of the individual subclasses.
For example, packages like \pkg{beachmat} can check if the current delayed layer is a no-op class and immediately move onto processing the seed.
}
