% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadArray.R
\name{loadArray}
\alias{loadArray}
\title{Load high-dimensional arrays}
\usage{
loadArray(info, project)
}
\arguments{
\item{info}{Named list containing the metadata for this array.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}. 
By default, this should be a string containing the path to a staging directory.}
}
\value{
A multi-dimensional object (usually a \linkS4class{DelayedMatrix}) containing the array data.
}
\description{
Default loading of arrays from on-disk formats, using the corresponding \code{\link{stageObject}} method. 
It should not be necessary for users to call this function manually.
}
\examples{
dir <- tempfile()
dir.create(dir)

arr <- array(rpois(10000, 10), c(50, 20, 10))
dimnames(arr) <- list(
   paste0("GENE_", seq_len(nrow(arr))),
   letters[1:20],
   NULL
)

path <- "whee"
info <- stageObject(arr, dir, path)
loadArray(info, project=dir)

}
\author{
Aaron Lun
}
