% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleMatrix.R
\name{MultiSampleMatrix}
\alias{MultiSampleMatrix}
\alias{MultiSampleMatrix-class}
\alias{MultiSampleMatrixSeed}
\alias{MultiSampleMatrixSeed-class}
\alias{DelayedArray,MultiSampleMatrixSeed-method}
\alias{listSamples}
\alias{extractSamples}
\title{Multi-sample matrix}
\description{
Implements a multi-sample matrix as a delayed \code{\link{cbind}} of per-sample DelayedMatrix objects,
but with some options for extracting and listing the constituent matrices for each sample.
This allows \code{\link{stageObject}} to save multiple samples without aggregating them into a single file.
}
\section{Constructors}{

\code{MultiSampleMatrixSeed(...)} accepts any number of named matrix-like objects and returns a MultiSampleMatrixSeed.
Each object corresponds to a block and should be named accordingly; names should be unique and non-empty.

\code{MultiSampleMatrix(...)} accepts any number of named matrix-like objects and returns a MultiSampleMatrix.
Alternatively, a single MultiSampleMatrixSeed may be provided.
}

\section{Functions}{

\code{listSamples(x)} will return a character vector of names of samples in a MultiSampleMatrix(Seed) object \code{x}.

\code{extractSamples(x)} will return a named list of matrix-like objects,
corresponding to the per-sample matrices used to construct the MultiSampleMatrix(Seed) object \code{x}.
}

\section{Comments on usage}{

The MultiSampleMatrixSeed is closely related to (and in fact, is a subclass of) the \linkS4class{DelayedAbind} class.
This means that we can leverage many of the \pkg{DelayedArray} methods for handling the delayed bind.
In theory, we could just use a DelayedAbind directly and save it with \pkg{chihaya} in \code{\link{stageObject}} (via \code{\link{preserveDelayedOperations}(TRUE)}).
However, this provides fewer opportunities for tracking and manipulating the samples.
It also saves the per-sample matrices into a single file, which eliminates possibilities for per-file deduplication and linking, e.g., with \code{\link{recycleHdf5Files}(TRUE)}.
}

\examples{
first <- Matrix::rsparsematrix(10, 10, 0.1)
second <- Matrix::rsparsematrix(10, 20, 0.1)
mat <- MultiSampleMatrix(foo = first, bar = second)
mat

listSamples(mat)
out <- extractSamples(mat)
lapply(out, dim)

}
\author{
Aaron Lun
}
