% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArray.R
\name{readArray}
\alias{readArray}
\title{Read high-dimensional arrays from disk}
\usage{
readArray(path, array.file.backed = TRUE, ...)
}
\arguments{
\item{info}{Named list containing the metadata for this array.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}. 
By default, this should be a string containing the path to a staging directory.}
}
\value{
A multi-dimensional object (usually a \linkS4class{DelayedMatrix}) containing the array data.
}
\description{
Read arrays from on-disk formats, using the corresponding \code{\link{stageObject}} method. 
It should not be necessary for users to call this function manually.
}
\examples{
arr <- array(rpois(10000, 10), c(50, 20, 10))
dimnames(arr) <- list(
   paste0("GENE_", seq_len(nrow(arr))),
   letters[1:20],
   NULL
)

dir <- tempfile()
saveObject(arr, dir)
readArray(dir)

}
\author{
Aaron Lun
}
