% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHDF5Array.R
\name{saveHDF5Array}
\alias{saveHDF5Array}
\alias{saveObject,HDF5Array-method}
\alias{saveObject,H5SparseMatrix-method}
\title{Save HDF5-backed arrays to disk}
\usage{
\S4method{saveObject}{HDF5Array}(x, path, array.dedup.session = NULL, array.dedup.action = NULL, ...)

\S4method{saveObject}{H5SparseMatrix}(x, path, array.dedup.session = NULL, array.dedup.action = NULL, ...)
}
\arguments{
\item{x}{A \linkS4class{DelayedArray} object.}

\item{path}{String containing a path to a directory in which to save \code{x}.}

\item{array.dedup.session, array.dedup.action}{Arguments controlling deduplication of \code{x}, see \code{?"\link{saveObject,array-method}"} for details.}

\item{...}{Further arguments, currently ignored.}
}
\value{
\code{x} is saved to \code{path} and \code{NULL} is invisibly returned.
}
\description{
Save \link[HDF5Array]{HDF5Array} ond \link[HDF5Array]{H5SparseMatrix} bjects to an on-disk representation.
}
\details{
These methods are provided solely to avoid \code{\link{saveObject}} dispatching to \code{\link{saveObject,DelayedArray-method}} upon encountering a HDF5-backed array.
If this dispatch occurs and \code{DelayedArray.preserve.ops=TRUE}, \code{x} will be saved as a DelayedArray on disk.
This is not incorrect per se but adds unnecessary files to the on-disk representation.
The methods listed here will avoid creating such files and save a bit of disk space.
}
\examples{
library(HDF5Array)
mat <- as(array(rpois(1000, 10), c(50, 20)), "HDF5Array")
dir <- tempfile()
saveObject(mat, dir)
list.files(dir)

}
\author{
Aaron Lun
}
