% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRawArraySeed.R
\name{createRawArraySeed}
\alias{createRawArraySeed}
\alias{.createRawArraySeed}
\alias{.extractArrayDimnames}
\title{Array loading utilities}
\usage{
.createRawArraySeed(info, path, names = TRUE)

.extractArrayDimnames(path, group, ndim)
}
\arguments{
\item{info}{A named list of metadata for this array.}

\item{path}{String containing the path to the file containing said array.}

\item{names}{Logical scalar indicating whether the seed should be annotated with dimnames (if available).}

\item{ndim}{Integer scalar specifying the number of dimensions.}
}
\value{
\code{.createRawArraySeed} returns a seed that can be used in the \code{\link{DelayedArray}} constructor.
For matrices, this will call \code{\link{.createRawMatrixSeed}}.

\code{.extractArrayDimnames} returns a list of character vectors or \code{NULL}, containing the dimnames.
}
\description{
Utilities for loading an array saved by \code{\link{stageObject}}.
}
\details{
For \code{.extractArrayDimnames}, \code{path} is expected to be a HDF5 file with a \code{names} group.
Each child of this group is a string dataset named after a (0-indexed) dimension, containing the names for that dimension.
}
\author{
Aaron Lun
}
