% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArray.R
\name{readArray}
\alias{readArray}
\title{Read a dense array from disk}
\usage{
readArray(path, array.file.backed = TRUE, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created by the \code{\link{saveObject}} method for a dense array.}

\item{array.file.backed}{Logical scalar indicating whether to return a file-backed S4 array class, or an ordinary R array in memory.}

\item{...}{Further arguments, ignored.}
}
\value{
A multi-dimensional array-like object, either a \linkS4class{DelayedArray} or an ordinary R array.
}
\description{
Read a dense high-dimensional array from its on-disk representation.
}
\examples{
arr <- array(rpois(10000, 10), c(50, 20, 10))
dimnames(arr) <- list(
   paste0("GENE_", seq_len(nrow(arr))),
   letters[1:20],
   NULL
)

dir <- tempfile()
saveObject(arr, dir)
readArray(dir)

}
\seealso{
\code{"\link{saveObject,array-method}"}, to create the directory and its contents.
}
\author{
Aaron Lun
}
