% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArray.R
\name{readArray}
\alias{readArray}
\alias{loadArray}
\title{Read a dense array from disk}
\usage{
readArray(path, metadata, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created by the \code{\link{saveObject}} method for a dense array.}

\item{metadata}{Named list of metadata for this object, see \code{\link{readObject}} for more details.}

\item{...}{Further arguments, ignored.}
}
\value{
A dense file-backed \linkS4class{ReloadedArray}.
}
\description{
Read a dense high-dimensional array from its on-disk representation.
This is usually not directly called by users, but is instead called by dispatch in \code{\link{readObject}}.
}
\examples{
arr <- array(rpois(10000, 10), c(50, 20, 10))
dimnames(arr) <- list(
   paste0("GENE_", seq_len(nrow(arr))),
   letters[1:20],
   NULL
)

dir <- tempfile()
saveObject(arr, dir)
readObject(dir)

}
\seealso{
\code{"\link{saveObject,array-method}"}, to create the directory and its contents.
}
\author{
Aaron Lun
}
