% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseMatrix.R
\name{readSparseMatrix}
\alias{readSparseMatrix}
\title{Read a sparse matrix from disk}
\usage{
readSparseMatrix(path, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created by the \code{\link{saveObject}} method for a spars matrix.}

\item{...}{Further arguments, ignored.}
}
\value{
A sparse \linkS4class{DelayedMatrix} object.
}
\description{
Read a sparse matrix from its on-disk representation.
}
\examples{
mat <- Matrix::rsparsematrix(100, 200, density=0.2)
rownames(mat) <- paste0("GENE_", seq_len(nrow(mat)))
dir <- tempfile()
saveObject(mat, dir)
readSparseMatrix(dir)

}
\seealso{
\code{"\link{saveObject,sparseMatrix-method}"}, to create the directory and its contents.
}
\author{
Aaron Lun
}
