% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDelayedArray.R
\name{saveDelayedArray}
\alias{saveDelayedArray}
\alias{saveObject,DelayedArray-method}
\title{Save DelayedArrays to disk}
\usage{
\S4method{saveObject}{DelayedArray}(x, path, delayedarray.preserve.ops = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{DelayedArray} object.}

\item{path}{String containing a path to a directory in which to save \code{x}.}

\item{delayedarray.preserve.ops}{Logical scalar indicating whether delayed operations should be preserved on-disk.}

\item{...}{Further arguments, ignored.}
}
\value{
\code{x} is saved to \code{path} and \code{NULL} is invisibly returned.
}
\description{
Save \link{DelayedArray} objects to their on-disk representation.
}
\examples{
mat <- Matrix::rsparsematrix(100, 200, density=0.2)
rownames(mat) <- paste0("GENE_", seq_len(nrow(mat)))
dmat <- DelayedArray::DelayedArray(mat) * 1

dir <- tempfile()
saveObject(dmat, dir)
list.files(dir)

}
\author{
Aaron Lun
}
