% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/storeDelayedObject.R
\name{storeDelayedObject}
\alias{storeDelayedObject}
\alias{reloadDelayedObject}
\alias{storeDelayedObject,ConstantArraySeed-method}
\alias{storeDelayedObject,DelayedAbind-method}
\alias{storeDelayedObject,ANY-method}
\alias{storeDelayedObject,DelayedAperm-method}
\alias{storeDelayedObject,DelayedNaryIsoOp-method}
\alias{storeDelayedObject,DelayedSetDimnames-method}
\alias{storeDelayedObject,DelayedSubassign-method}
\alias{storeDelayedObject,DelayedSubset-method}
\alias{storeDelayedObject,DelayedUnaryIsoOpStack-method}
\alias{storeDelayedObject,DelayedUnaryIsoOpWithArgs-method}
\alias{storeDelayedObject,SVT_SparseMatrix-method}
\alias{storeDelayedObject,array-method}
\alias{storeDelayedObject,denseMatrix-method}
\alias{storeDelayedObject,sparseMatrix-method}
\alias{altStoreDelayedObject}
\alias{altStoreDelayedObjectFunction}
\alias{altReloadDelayedObject}
\alias{altReloadDelayedObjectFunction}
\title{Store/reload operations in a DelayedArray}
\usage{
storeDelayedObject(x, handle, name, ...)

reloadDelayedObject(handle, name, version = package_version("1.1"), ...)

altStoreDelayedObjectFunction(store)

altStoreDelayedObject(...)

altReloadDelayedObjectFunction(reload)

altReloadDelayedObject(...)
}
\arguments{
\item{x}{Any of the delayed operation classes from \pkg{DelayedArray}.}

\item{handle}{An \pkg{rhdf5} handle of a HDF5 file to save into (for \code{storeDelayedObject}) or load from (for \code{reloadDelayedObject}).}

\item{name}{String containing the name of the group in \code{file} to save into (for \code{storeDelayedObject}) or load from (for \code{reloadDelayedObject}).}

\item{...}{For \code{storeDelayedObject} and \code{reloadDelayedObject}, additional arguments to be passed to specific methods.

For \code{altStoreDelayedObject} and \code{altReloadDelayedObject}, arguments to be passed to the alternative functions.}

\item{version}{Package version of the \pkg{chihaya} format to use when loading.
This should be retrieved from the attributes of the outermost group, typically by \code{readDelayedArray}.}

\item{store}{Function (typically a generic) to store delayed operations to file.
This should accept the same arguments as \code{storeDelayedObject}.}

\item{reload}{Function to reload delayed operations from file.
This should accept the same arguments as \code{reloadDelayedObject}.}
}
\value{
For \code{storeDelayedObject} and \code{altStoreDelayedObject}, the contents of \code{x} are saved to \code{file}, and \code{NULL} is invisibly returned.

For \code{reloadDelayedObject} and \code{altReloadDelayedObject}, a delayed operation or \linkS4class{DelayedArray} is returned.

For \code{altStoreDelayedObjectFunction}, the current store function is returned if \code{store} is missing.
Otherwise, \code{store} is set as the current store function and the previous store function is returned.

For \code{altReloadDelayedObjectFunction}, the current reload function is returned if \code{reload} is missing.
Otherwise, \code{reload} is set as the current reload function and the previous reload function is returned.
}
\description{
Store or reload the delayed operations of a \linkS4class{DelayedArray} in an existing HDF5 file.
}
\section{Customization}{

Application developers can customize the process of storing/reloading delayed operations by specifying alternative functions in \code{altReloadDelayedObjectFunction} and \code{altStoreDelayedObjectFunction}.
For example, if we want to preserve all delayed operations except for \linkS4class{DelayedSubset},
we could replace \code{storeDelayedObject} with an \code{altStoreDelayedObject} that realizes any DelayedSubset instance into an ordinary matrix.
This is analogous to the overrides for \code{\link{altReadObject}} and \code{\link{altSaveObject}}.
}

\examples{
library(DelayedArray)
X <- DelayedArray(matrix(runif(100), ncol=20))
Y <- cbind(X, DelayedArray::ConstantArray(value=50, c(5, 10)))

library(rhdf5)
temp <- tempfile()
dir.create(temp)

fpath <- file.path(temp, "foo.h5")
fhandle <- H5Fcreate(fpath)
storeDelayedObject(Y@seed, fhandle, "YAY")
rhdf5::h5ls(fhandle)
H5Fclose(fhandle)

fhandle <- H5Fopen(fpath, "H5F_ACC_RDONLY")
reloadDelayedObject(fhandle, "YAY")
H5Fclose(fhandle)

}
\seealso{
\code{\link{saveObject,DelayedArray-method}} and \code{\link{readDelayedArray}}, where these methods are used.
}
\author{
Aaron Lun
}
