% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBeRealizedArray.R
\name{ToBeRealizedArraySeed}
\alias{ToBeRealizedArraySeed}
\alias{ToBeRealizedArraySeed-class}
\alias{ToBeRealizedArray-class}
\alias{ToBeRealizedMatrix-class}
\alias{DelayedArray,ToBeRealizedArraySeed-method}
\alias{matrixClass,ToBeRealizedArray-method}
\alias{coerce,ToBeRealizedArray,ToBeRealizedMatrix-method}
\alias{coerce,ToBeRealizedMatrix,ToBeRealizedArray-method}
\alias{saveObject,ToBeRealizedArray-method}
\alias{ToBeRealizedArray}
\title{To-be-realized array}
\usage{
ToBeRealizedArraySeed(seed)

ToBeRealizedArray(seed)

\S4method{saveObject}{ToBeRealizedArray}(x, path, array.dedup.session = NULL, array.dedup.action = NULL, ...)
}
\arguments{
\item{seed}{Seed representing an array-like object, typically a \link[DelayedArray]{DelayedArray}.}

\item{x}{A \linkS4class{ToBeRealizedArray} object.}

\item{path}{String containing a path to a directory in which to save \code{x}.}

\item{array.dedup.session, array.dedup.action}{Arguments controlling deduplication of \code{x}, see \code{?"\link{saveObject,array-method}"} for details.}

\item{...}{Further arguments to pass to the relevant \code{\link{saveObject}} methods for saving dense arrays and sparse matrices.}
}
\value{
For the constructors, an instance of the \linkS4class{ToBeRealizedArraySeed} or \linkS4class{ToBeRealizedArray}.

For \code{saveObject}, \code{x} is saved to \code{path} and \code{NULL} is returned.
}
\description{
A DelayedArray layer that indicates that its seed is to be realized by \code{\link{saveObject,DelayedArray-method}}.
This is intended for fine-grained control over which delayed operations are to be preserved in an object with multiple arrays, e.g., a SummarizedExperiment.
}
\details{
The ToBeRealizedArraySeed is a \linkS4class{DelayedUnaryIsoOp} subclass that will just forward all operations to the underlying \code{seed}.
Its purpose is to indicate to \code{\link{saveObject}} that any delayed operations in its \code{seed} should not be preserved, even if \code{DelayedArray.preserve.ops=TRUE}.
This is occasionally useful when saving a complicated object with many arrays like a SummarizedExperiment,
where we want to preserve delayed operations in some arrays while realizing operations in others.
}
\examples{
library(DelayedArray)
x1 <- matrix(runif(100), 5, 20)
x2 <- matrix(runif(250), 5, 50)
x <- cbind(DelayedArray(x1), DelayedArray(x2))
y0 <- matrix(runif(300), 10, 30)
y <- log1p(x)
big.object <- list(foo=x, bar=y)

# Imagine we want to save 'big.object', preserving the delayed log1p
# for 'y' but realizing the delayed cbind for 'x'. This would not be
# possible with just DelayedArray.preserve.ops=TRUE as both sets of
# delayed operations would be preserved:
tmp <- tempfile()
saveObject(big.object, tmp, DelayedArray.preserve.ops=TRUE)
list.files(tmp, recursive=TRUE)
readObjectFile(file.path(tmp, "other_contents", "0"))$type
readObjectFile(file.path(tmp, "other_contents", "1"))$type

# Instead we wrap 'x' in a ToBeRealizedArray, which forces its realization.
big.object2 <- list(foo=ToBeRealizedArray(x), bar=y)
tmp2 <- tempfile()
saveObject(big.object2, tmp2, DelayedArray.preserve.ops=TRUE)
list.files(tmp2, recursive=TRUE)
readObjectFile(file.path(tmp2, "other_contents", "0"))$type
readObjectFile(file.path(tmp2, "other_contents", "1"))$type

}
\author{
Aaron Lun
}
