% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperArraySeed.R
\name{WrapperArraySeed}
\alias{WrapperArraySeed}
\alias{WrapperArraySeed-class}
\alias{WrapperArray-class}
\alias{WrapperArray}
\alias{loadWrapperArray}
\title{DelayedArray wrapper seed}
\description{
Deprecated, use \link[DelayedArray]{DelayedUnaryIsoOp} instead.
}
